requirements.txt
Flask[async]==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-json-logger==2.0.7
python-dotenv==1.0.0
ddtrace==2.3.0
aiohttp==3.9.1
asyncpg==0.29.0
httpx==0.25.2
